#2878: split test into individual dbmax files
