--- test/live_server/2_database/pytest_svr_dbmax.py
@@ -1,378 +1,0 @@
-'''
-
-This file will test specific database limitations. More generally, it will test
-conditions when a user chooses to store a new collection of documents, or when
-users elect to append additional documents, into a preexisting collection. The
-following svr cases for both anonymous, and authenticated users will be tested:
-
-  - users have exceeded maximum collections
-    - anon: oldest collection will be removed, from corresponding database(s),
-            prior to saving the new proposed collection
-    - auth: new collection will not save, unless the user chooses to manually
-            delete a previous collection of their choice
-
-  - users have exceeded maximum documents in a collections
-    - anon: new documents will not save, in the proposed collection
-    - auth: new documents will not save, unless the user chooses to manually
-            delete a previous document, in the proposed collection
-
-Note: the 'pytest' instances can further be reviewed:
-
-    - https://pytest-flask.readthedocs.io/en/latest
-    - http://docs.pytest.org/en/latest/usage.html
-
-'''
-
-import json
-import os.path
-from flask import current_app, url_for
-
-
-def get_sample_json(jsonfile, model_type):
-    '''
-
-    Get a sample json dataset.
-
-    '''
-
-    # local variables
-    root = current_app.config.get('ROOT')
-
-    # open file
-    json_dataset = None
-
-    with open(
-        os.path.join(
-            root,
-            'interface',
-            'static',
-            'data',
-            'json',
-            'programmatic_interface',
-            model_type,
-            'dataset_url',
-            jsonfile
-        ),
-        'r'
-    ) as json_file:
-        json_dataset = json.load(json_file)
-
-    return json_dataset
-
-
-def test_max_collections_anon(client, live_server):
-    '''
-
-    This method will test when an anonymous user, has exceeded the maximum
-    allowed collection. In such a case, the oldest collection will be removed,
-    to allow the proposed collection to be stored in the corresponding
-    database(s).
-
-    '''
-
-    @live_server.app.route('/load-data')
-    def load_data():
-        return url_for('name.load_data', _external=True)
-
-    @live_server.app.route('/collection-count')
-    def collection_count():
-        return url_for('name.collection_count', _external=True)
-
-    @live_server.app.route('/document-count')
-    def document_count():
-        return url_for('name.document_count', _external=True)
-
-    @live_server.app.route('/remove-collection')
-    def remove_collection():
-        return url_for('name.remove_collection', _external=True)
-
-    live_server.start()
-
-    # local variables
-    uid = 0
-    max_collection = current_app.config.get('MAXCOL_ANON')
-    max_document = current_app.config.get('MAXDOC_ANON')
-
-    # save max collection
-    for i in range(max_collection):
-        dataset = get_sample_json('svr-data-new.json', 'svr')
-        dataset['properties']['collection'] = 'collection--pytest-svr--' + str(i)
-
-        res = client.post(
-            load_data(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps(dataset)
-        )
-
-        # assertion checks
-        assert res.status_code == 200
-        assert res.json['status'] == 0
-
-        res = client.post(
-            document_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-        assert res.json['count'] == i + 1
-
-        res = client.post(
-            document_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({'uid': uid})
-        )
-
-        assert res.status_code == 200
-        assert res['count'] == i + 1
-
-    # save max collection + 1
-    dataset = get_sample_json('svr-data-new.json', 'svr')
-    dataset['properties']['collection'] = 'collection--pytest-svr--' + str(i + 1)
-
-    res = client.post(
-        load_data(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps(dataset)
-    )
-
-    assert res.status_code == 200
-    assert res.json['status'] == 0
-
-    res = client.post(
-        collection_count(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps({'uid': uid})
-    )
-
-    assert res.status_code == 200
-    assert res['count'] == max_collection - 1
-
-    res = client.post(
-        document_count(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps({
-            'collection': 'collection--pytest-svr--' + str(i),
-        })
-    )
-
-    assert res.status_code == 200
-    assert res['count'] > 0
-    assert res['count'] < max_document
-
-    res = client.post(
-        document_count(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps({
-            'collection': 'collection--pytest-svr--' + str(i + 1),
-        })
-    )
-
-    assert res.status_code == 200
-    assert res['count'] == 0
-
-    # drop all collections and related entities
-    for i in range(max_collection):
-        res = client.post(
-            remove_collection(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-
-        res = client.post(
-            remove_collection(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'uid': uid,
-                'type': 'entity',
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-
-        res = client.post(
-            document_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-        assert res['count'] == 0
-
-        res = client.post(
-            collection_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({'uid': uid})
-        )
-
-        assert res.status_code == 200
-        assert res['count'] == 0
-
-
-def test_max_collections_auth(client, live_server):
-    '''
-
-    This method will test when an authenticated user, has exceeded the maximum
-    allowed collection. In such a case, the proposed collection will not to be
-    stored in the corresponding database(s). Instead, the authenticated user,
-    will need to manually delete a chosen collection, using constructs of the
-    corresponding web-interface, or exposed programmatic-api, to allow the
-    proposed collection to be stored in the corresponding database(s).
-
-    '''
-
-    @live_server.app.route('/load-data')
-    def load_data():
-        return url_for('name.load_data', _external=True)
-
-    @live_server.app.route('/collection-count')
-    def collection_count():
-        return url_for('name.collection_count', _external=True)
-
-    @live_server.app.route('/document-count')
-    def document_count():
-        return url_for('name.document_count', _external=True)
-
-    @live_server.app.route('/remove-collection')
-    def remove_collection():
-        return url_for('name.remove_collection', _external=True)
-
-    live_server.start()
-
-    # local variables
-    uid = 1
-    max_collection = current_app.config.get('MAXCOL_AUTH')
-    max_document = current_app.config.get('MAXDOC_AUTH')
-
-    # save max collection
-    for i in range(max_collection):
-        dataset = get_sample_json('svr-data-new.json', 'svr')
-        dataset['properties']['collection'] = 'collection--pytest-svr--' + str(i)
-
-        res = client.post(
-            load_data(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps(dataset)
-        )
-
-        # assertion checks
-        assert res.status_code == 200
-        assert res.json['status'] == 0
-
-        res = client.post(
-            document_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-        assert res.json['count'] == i + 1
-
-        res = client.post(
-            document_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({'uid': uid})
-        )
-
-        assert res.status_code == 200
-        assert res['count'] == i + 1
-
-    # save max collection + 1
-    dataset = get_sample_json('svr-data-new.json', 'svr')
-    dataset['properties']['collection'] = 'collection--pytest-svr--' + str(i + 1)
-
-    res = client.post(
-        load_data(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps(dataset)
-    )
-
-    assert res.status_code == 200
-    assert res.json['status'] == 0
-
-    res = client.post(
-        collection_count(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps({'uid': uid})
-    )
-
-    assert res.status_code == 200
-    assert res['count'] == max_collection - 1
-
-    res = client.post(
-        document_count(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps({
-            'collection': 'collection--pytest-svr--' + str(i),
-        })
-    )
-
-    assert res.status_code == 200
-    assert res['count'] > 0
-    assert res['count'] < max_document
-
-    res = client.post(
-        document_count(),
-        headers={'Content-Type': 'application/json'},
-        data=json.dumps({
-            'collection': 'collection--pytest-svr--' + str(i + 1),
-        })
-    )
-
-    assert res.status_code == 200
-    assert res['count'] == 0
-
-    # drop all collections and related entities
-    for i in range(max_collection):
-        res = client.post(
-            remove_collection(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-
-        res = client.post(
-            remove_collection(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'uid': uid,
-                'type': 'entity',
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-
-        res = client.post(
-            document_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({
-                'collection': 'collection--pytest-svr--' + str(i),
-            })
-        )
-
-        assert res.status_code == 200
-        assert res['count'] == 0
-
-        res = client.post(
-            collection_count(),
-            headers={'Content-Type': 'application/json'},
-            data=json.dumps({'uid': uid})
-        )
-
-        assert res.status_code == 200
-        assert res['count'] == 0
